pyproject.toml
[project]
name = "prompt-anywhere"
version = "0.1.0"
description = "Global hotkey screenshot → Gemini CLI analyzer"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "pynput>=1.7.6",
    "Pillow>=10.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
]

[project.scripts]
prompt-anywhere = "prompt_anywhere.ui.app:main"
prompt-anywhere-host = "prompt_anywhere.host.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
